fix(db): separate current state of the releases from releases in history
Ref: SRX-2HNB0H
