+++ pkg/db/db.go
@@ -514,773 +514,6 @@ func (h *DBHandler) DBReadEslEventLaterThan(ctx context.Context, tx *sql.Tx, esl
 	return row, nil
 }
 
-// RELEASES
-// Releases work a bit different, because they are already immutable.
-// We still store the eslVersion for consistency with other tables,
-// but technically it's not required here.
-
-type DBReleaseMetaData struct {
-	SourceAuthor    string
-	SourceCommitId  string
-	SourceMessage   string
-	DisplayVersion  string
-	UndeployVersion bool
-	IsMinor         bool
-	CiLink          string
-	IsPrepublish    bool
-}
-
-type DBReleaseManifests struct {
-	Manifests map[string]string
-}
-
-type DBReleaseWithMetaData struct {
-	ReleaseNumber uint64
-	Created       time.Time
-	App           string
-	Manifests     DBReleaseManifests
-	Metadata      DBReleaseMetaData
-	Deleted       bool
-	Environments  []string
-}
-
-type DBAllReleaseMetaData struct {
-	Releases []int64
-}
-type DBAllReleasesWithMetaData struct {
-	EslVersion EslVersion
-	Created    time.Time
-	App        string
-	Metadata   DBAllReleaseMetaData
-}
-
-func (h *DBHandler) DBSelectAnyRelease(ctx context.Context, tx *sql.Tx, ignorePrepublishes bool) (*DBReleaseWithMetaData, error) {
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, appName, metadata, releaseVersion, deleted, environments" +
-			" FROM releases " +
-			" LIMIT 1;"))
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAnyRelease")
-	defer span.Finish()
-	span.SetTag("query", selectQuery)
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-	)
-	processedRows, err := h.processReleaseRows(ctx, err, rows, ignorePrepublishes, false)
-	if err != nil {
-		return nil, err
-	}
-	if len(processedRows) == 0 {
-		return nil, nil
-	}
-	return processedRows[0], nil
-}
-
-func (h *DBHandler) DBSelectReleasesWithoutEnvironments(ctx context.Context, tx *sql.Tx) ([]*DBReleaseWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleasesWithoutEnvironments")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(`
-	SELECT DISTINCT
-		releases.created,
-		releases.appName,
-		releases.metadata,
-		releases.manifests,
-		releases.releaseVersion,
-		releases.deleted,
-		releases.environments
-	FROM (
-		SELECT
-			MAX(version) AS latestVersion,
-			appname,
-			releaseversion
-		FROM
-			"releases"
-		GROUP BY
-			appname, releaseversion) AS latest
-	JOIN
-		releases AS releases
-	ON
-		latest.latestVersion=releases.version
-		AND latest.releaseVersion=releases.releaseVersion
-		AND latest.appname=releases.appname
-	WHERE COALESCE(environments, '') = '' AND COALESCE(manifests, '') != ''
-	LIMIT 100;
-	`)
-	rows, err := tx.QueryContext(ctx, selectQuery)
-	processedRows, err := h.processReleaseRows(ctx, err, rows, true, true)
-	if err != nil {
-		return nil, err
-	}
-	return processedRows, nil
-}
-
-func (h *DBHandler) DBSelectReleasesByVersions(ctx context.Context, tx *sql.Tx, app string, releaseVersions []uint64, ignorePrePublishes bool) ([]*DBReleaseWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleasesByVersions")
-	defer span.Finish()
-
-	if len(releaseVersions) == 0 {
-		return []*DBReleaseWithMetaData{}, nil
-	}
-	repeatedQuestionMarks := strings.Repeat(",?", len(releaseVersions)-1)
-	selectQuery := h.AdaptQuery(`
-
-	SELECT DISTINCT
-		releases.created,
-		releases.appName,
-		releases.metadata,
-		releases.releaseVersion,
-		releases.deleted,
-		releases.environments
-	FROM 
-		( 	SELECT
-				MAX(version) AS latestVersion,
-				appname,
-				releaseversion
-			FROM
-				"releases"
-    		WHERE
-				appname=?
-          	AND
-				releaseversion IN (?` + repeatedQuestionMarks + `)
-		GROUP BY
-			appname, releaseversion
-      ) AS latest
-	JOIN
-		releases AS releases
-	ON
-		latest.latestVersion=releases.version
-		AND latest.releaseVersion=releases.releaseVersion
-		AND latest.appname=releases.appname
-	LIMIT ?
-	`)
-
-	span.SetTag("query", selectQuery)
-
-	args := []any{}
-	args = append(args, app)
-	for _, version := range releaseVersions {
-		args = append(args, version)
-	}
-	args = append(args, uint64(len(releaseVersions)))
-
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		args...,
-	)
-
-	processedRows, err := h.processReleaseRows(ctx, err, rows, ignorePrePublishes, false)
-	if err != nil {
-		return nil, err
-	}
-	if len(processedRows) == 0 {
-		return nil, nil
-	}
-	return processedRows, nil
-}
-
-func (h *DBHandler) DBSelectReleaseByVersion(ctx context.Context, tx *sql.Tx, app string, releaseVersion uint64, ignorePrepublishes bool) (*DBReleaseWithMetaData, error) {
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, appName, metadata, manifests, releaseVersion, deleted, environments " +
-			" FROM releases " +
-			" WHERE appName=? AND releaseVersion=?" +
-			" ORDER BY version DESC " +
-			" LIMIT 1;"))
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleaseByVersion")
-	defer span.Finish()
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		app,
-		releaseVersion,
-	)
-
-	processedRows, err := h.processReleaseRows(ctx, err, rows, ignorePrepublishes, true)
-	if err != nil {
-		return nil, err
-	}
-	if len(processedRows) == 0 {
-		return nil, nil
-	}
-	return processedRows[0], nil
-}
-
-func (h *DBHandler) DBSelectReleaseByVersionAtTimestamp(ctx context.Context, tx *sql.Tx, app string, releaseVersion uint64, ignorePrepublishes bool, ts time.Time) (*DBReleaseWithMetaData, error) {
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT created, appName, metadata, manifests, releaseVersion, deleted, environments " +
-			" FROM releases " +
-			" WHERE appName=? AND releaseVersion=? AND created <= (?)" +
-			" ORDER BY version DESC " +
-			" LIMIT 1;"))
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleaseByVersion")
-	defer span.Finish()
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		app,
-		releaseVersion,
-		ts,
-	)
-
-	processedRows, err := h.processReleaseRows(ctx, err, rows, ignorePrepublishes, true)
-	if err != nil {
-		return nil, err
-	}
-	if len(processedRows) == 0 {
-		return nil, nil
-	}
-	return processedRows[0], nil
-}
-
-func (h *DBHandler) DBSelectAllManifestsForAllReleases(ctx context.Context, tx *sql.Tx) (map[string]map[uint64][]string, error) {
-	selectQuery := h.AdaptQuery(
-		`
-	SELECT DISTINCT
-		releases.appname,
-		releases.releaseVersion,
-		releases.environments
-	FROM (
-		SELECT
-			MAX(version) AS latestRelease,
-			appname,
-			releaseversion
-		FROM
-			"releases"
-		GROUP BY
-			appname, releaseversion) AS latest
-	JOIN
-		releases AS releases
-	ON
-		latest.latestRelease=releases.version
-		AND latest.releaseVersion=releases.releaseVersion
-		AND latest.appname=releases.appname;
-`)
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllManifestsForAllReleases")
-	defer span.Finish()
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-	)
-
-	return h.processReleaseManifestRows(ctx, err, rows)
-}
-
-func (h *DBHandler) processReleaseManifestRows(ctx context.Context, err error, rows *sql.Rows) (map[string]map[uint64][]string, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "processReleaseManifestRows")
-	defer span.Finish()
-
-	if err != nil {
-		return nil, fmt.Errorf("could not query releases table from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("releases: row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	var result = make(map[string]map[uint64][]string)
-	for rows.Next() {
-		var environmentsStr sql.NullString
-		var appName string
-		var releaseVersion uint64
-		err := rows.Scan(&appName, &releaseVersion, &environmentsStr)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning releases row from DB. Error: %w\n", err)
-		}
-		environments := make([]string, 0)
-		if environmentsStr.Valid && environmentsStr.String != "" {
-			err = json.Unmarshal(([]byte)(environmentsStr.String), &environments)
-			if err != nil {
-				return nil, fmt.Errorf("Error during json unmarshal of environments for releases. Error: %w. Data: %s\n", err, environmentsStr.String)
-			}
-		}
-		if _, exists := result[appName]; !exists {
-			result[appName] = make(map[uint64][]string)
-		}
-		result[appName][releaseVersion] = environments
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-// DBSelectReleasesByAppLatestEslVersion returns the latest eslversion
-// for each release of an app. It includes deleted releases and loads manifests.
-func (h *DBHandler) DBSelectReleasesByAppLatestEslVersion(ctx context.Context, tx *sql.Tx, app string, ignorePrepublishes bool) ([]*DBReleaseWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleasesByAppLatestEslVersion")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(
-		`SELECT
-			releases.created,
-			releases.appName,
-			releases.metadata,
-			releases.manifests,
-			releases.releaseVersion,
-			releases.deleted,
-			releases.environments
-		FROM (
-			SELECT
-				MAX(version) AS latestVersion,
-				appname,
-				releaseversion
-			FROM
-				releases
-			WHERE 
-				appname=?
-			GROUP BY
-				appname, releaseversion
-		) as currentReleases
-		JOIN
-		  releases
-		ON 
-			currentReleases.appname 			= releases.appname
-		AND
-			currentReleases.latestVersion = releases.version
-		AND
-			currentReleases.releaseversion 	= releases.releaseversion
-		ORDER BY currentReleases.releaseversion DESC;`,
-	)
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		app,
-	)
-
-	return h.processReleaseRows(ctx, err, rows, ignorePrepublishes, true)
-}
-
-func (h *DBHandler) DBSelectReleasesByAppOrderedByEslVersion(ctx context.Context, tx *sql.Tx, app string, ignorePrepublishes bool) (*DBReleaseWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectReleasesByAppOrderedByEslVersion")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(`
-SELECT
-	releases.created,
-	releases.appName,
-	releases.metadata,
-	releases.releaseVersion,
-	releases.deleted,
-	releases.environments
-FROM (
- 	SELECT max(version) as latestVersion, appName, releaseVersion
- 	FROM releases
- 	WHERE appName=?
- 	GROUP BY appName, releaseVersion
- 	) as c
-JOIN releases
-    ON  c.appName= releases.appName
-    AND c.latestVersion = releases.version
-    AND c.releaseVersion = releases.releaseVersion
-ORDER BY c.releaseVersion DESC
-LIMIT 1;
-`,
-	)
-	span.SetTag("query", selectQuery)
-	span.SetTag("appName", app)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		app,
-	)
-
-	result, err := h.processReleaseRows(ctx, err, rows, ignorePrepublishes, false)
-	span.SetTag("resultCount", len(result))
-	if err != nil {
-		span.Finish(tracer.WithError(err))
-		return nil, err
-	}
-	if result == nil {
-		return nil, nil
-	}
-	return result[0], nil
-}
-
-func (h *DBHandler) DBSelectAllReleasesOfApp(ctx context.Context, tx *sql.Tx, app string) (*DBAllReleasesWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllReleasesOfApp")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(fmt.Sprintf(
-		"SELECT eslVersion, created, appName, metadata " +
-			" FROM all_releases " +
-			" WHERE appName=?" +
-			" ORDER BY eslVersion DESC " +
-			" LIMIT 1;"))
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-		app,
-	)
-	return h.processAllReleasesRow(ctx, err, rows)
-}
-
-func (h *DBHandler) DBSelectAllReleasesOfAllApps(ctx context.Context, tx *sql.Tx) (map[string][]int64, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBSelectAllReleasesOfAllApps")
-	defer span.Finish()
-	selectQuery := h.AdaptQuery(`SELECT
-	all_releases.appname,
-		all_releases.metadata
-	FROM (
-		SELECT
-	MAX(eslVersion) AS latest,
-		appname
-	FROM
-		all_releases
-	GROUP BY
-		appname) AS latest
-	JOIN
-		all_releases AS all_releases
-	ON
-		latest.latest=all_releases.eslVersion
-		AND latest.appname=all_releases.appname;`)
-
-	span.SetTag("query", selectQuery)
-	rows, err := tx.QueryContext(
-		ctx,
-		selectQuery,
-	)
-	return h.processAllReleasesForAllAppsRow(ctx, err, rows)
-}
-
-func (h *DBHandler) processAllReleasesForAllAppsRow(ctx context.Context, err error, rows *sql.Rows) (map[string][]int64, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "processAllReleasesForAllAppsRow")
-	defer span.Finish()
-
-	if err != nil {
-		return nil, fmt.Errorf("could not query releases table from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("releases: row could not be closed: %v", err)
-		}
-	}(rows)
-
-	var result = make(map[string][]int64)
-	for rows.Next() {
-		var appName string
-		var jsonReleases string
-
-		err := rows.Scan(&appName, &jsonReleases)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning releases row from DB. Error: %w\n", err)
-		}
-		var metaData = DBAllReleaseMetaData{
-			Releases: []int64{},
-		}
-
-		err = json.Unmarshal(([]byte)(jsonReleases), &metaData)
-		if err != nil {
-			return nil, fmt.Errorf("Error during json unmarshal of releases. Error: %w. Data: %s\n", err, jsonReleases)
-		}
-		result[appName] = metaData.Releases
-	}
-
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (h *DBHandler) processAllReleasesRow(ctx context.Context, err error, rows *sql.Rows) (*DBAllReleasesWithMetaData, error) {
-	if err != nil {
-		return nil, fmt.Errorf("could not query releases table from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("releases: row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	var row = &DBAllReleasesWithMetaData{}
-	if rows.Next() {
-		var metadataStr string
-		err := rows.Scan(&row.EslVersion, &row.Created, &row.App, &metadataStr)
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning releases row from DB. Error: %w\n", err)
-		}
-		var metaData = DBAllReleaseMetaData{
-			Releases: []int64{},
-		}
-		err = json.Unmarshal(([]byte)(metadataStr), &metaData)
-		if err != nil {
-			return nil, fmt.Errorf("Error during json unmarshal of releases. Error: %w. Data: %s\n", err, metadataStr)
-		}
-		row.Metadata = metaData
-	} else {
-		row = nil
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return row, nil
-}
-
-func (h *DBHandler) processReleaseRows(ctx context.Context, err error, rows *sql.Rows, ignorePrepublishes bool, withManifests bool) ([]*DBReleaseWithMetaData, error) {
-	span, ctx := tracer.StartSpanFromContext(ctx, "processReleaseRows")
-	defer span.Finish()
-
-	if err != nil {
-		return nil, fmt.Errorf("could not query releases table from DB. Error: %w\n", err)
-	}
-	defer func(rows *sql.Rows) {
-		err := rows.Close()
-		if err != nil {
-			logger.FromContext(ctx).Sugar().Warnf("releases: row could not be closed: %v", err)
-		}
-	}(rows)
-	//exhaustruct:ignore
-	var result []*DBReleaseWithMetaData
-	var lastSeenRelease uint64 = 0
-	for rows.Next() {
-		//exhaustruct:ignore
-		var row = &DBReleaseWithMetaData{}
-		var metadataStr string
-		var manifestStr string
-		var environmentsStr sql.NullString
-		var err error
-		if withManifests {
-			err = rows.Scan(&row.Created, &row.App, &metadataStr, &manifestStr, &row.ReleaseNumber, &row.Deleted, &environmentsStr)
-		} else {
-			err = rows.Scan(&row.Created, &row.App, &metadataStr /*manifests*/, &row.ReleaseNumber, &row.Deleted, &environmentsStr)
-		}
-		if err != nil {
-			if errors.Is(err, sql.ErrNoRows) {
-				return nil, nil
-			}
-			return nil, fmt.Errorf("Error scanning releases row from DB withManifests=%v. Error: %w\n", withManifests, err)
-		}
-		if row.ReleaseNumber != lastSeenRelease {
-			lastSeenRelease = row.ReleaseNumber
-		} else {
-			continue
-		}
-		// handle meta data
-		var metaData = DBReleaseMetaData{
-			SourceAuthor:    "",
-			SourceCommitId:  "",
-			SourceMessage:   "",
-			DisplayVersion:  "",
-			UndeployVersion: false,
-			IsMinor:         false,
-			CiLink:          "",
-			IsPrepublish:    false,
-		}
-		err = json.Unmarshal(([]byte)(metadataStr), &metaData)
-		if err != nil {
-			return nil, fmt.Errorf("Error during json unmarshal of metadata for releases. Error: %w. Data: %s\n", err, metadataStr)
-		}
-		row.Metadata = metaData
-
-		// handle manifests
-		var manifestData = DBReleaseManifests{
-			Manifests: map[string]string{},
-		}
-		if withManifests {
-			err = json.Unmarshal(([]byte)(manifestStr), &manifestData)
-			if err != nil {
-				return nil, fmt.Errorf("Error during json unmarshal of manifests for releases. Error: %w. Data: %s\n", err, metadataStr)
-			}
-		}
-		row.Manifests = manifestData
-		environments := make([]string, 0)
-		if environmentsStr.Valid && environmentsStr.String != "" {
-			err = json.Unmarshal(([]byte)(environmentsStr.String), &environments)
-			if err != nil {
-				return nil, fmt.Errorf("Error during json unmarshal of environments for releases. Error: %w. Data: %s\n", err, environmentsStr.String)
-			}
-		}
-		row.Environments = environments
-		if ignorePrepublishes && row.Metadata.IsPrepublish {
-			continue
-		}
-		result = append(result, row)
-	}
-	err = closeRows(rows)
-	if err != nil {
-		return nil, err
-	}
-	return result, nil
-}
-
-func (h *DBHandler) DBInsertRelease(ctx context.Context, transaction *sql.Tx, release DBReleaseWithMetaData) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBInsertRelease")
-	defer span.Finish()
-	metadataJson, err := json.Marshal(release.Metadata)
-	if err != nil {
-		return fmt.Errorf("insert release: could not marshal json data: %w", err)
-	}
-	insertQuery := h.AdaptQuery(
-		"INSERT INTO releases (created, releaseVersion, appName, manifests, metadata, deleted, environments)  VALUES (?, ?, ?, ?, ?, ?, ?);",
-	)
-	span.SetTag("query", insertQuery)
-	manifestJson, err := json.Marshal(release.Manifests)
-	if err != nil {
-		return fmt.Errorf("could not marshal json data: %w", err)
-	}
-
-	envs := make([]string, 0)
-	for env := range release.Manifests.Manifests {
-		envs = append(envs, env)
-	}
-	release.Environments = envs
-	slices.Sort(release.Environments) // we don't really *need* the sorting, it's just for convenience
-	environmentStr, err := json.Marshal(release.Environments)
-	if err != nil {
-		return fmt.Errorf("could not marshal release environments: %w", err)
-	}
-
-	now, err := h.DBReadTransactionTimestamp(ctx, transaction)
-	if err != nil {
-		return fmt.Errorf("DBInsertRelease unable to get transaction timestamp: %w", err)
-	}
-	_, err = transaction.Exec(
-		insertQuery,
-		*now,
-		release.ReleaseNumber,
-		release.App,
-		manifestJson,
-		metadataJson,
-		release.Deleted,
-		environmentStr,
-	)
-	if err != nil {
-		return fmt.Errorf(
-			"could not insert release for app '%s' and version '%v' into DB. Error: %w\n",
-			release.App,
-			release.ReleaseNumber,
-			err)
-	}
-
-	logger.FromContext(ctx).Sugar().Infof(
-		"inserted release: app '%s' and version '%v'",
-		release.App,
-		release.ReleaseNumber)
-	return nil
-}
-
-func (h *DBHandler) DBDeleteReleaseFromAllReleases(ctx context.Context, transaction *sql.Tx, application string, releaseToDelete uint64) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBDeleteReleaseFromAllReleases")
-	defer span.Finish()
-
-	allReleases, err := h.DBSelectAllReleasesOfApp(ctx, transaction, application)
-	if err != nil {
-		return err
-	}
-	if allReleases == nil {
-		logger.FromContext(ctx).Sugar().Infof("Could not find release '%d' for appliation '%s' to delete. No releases found.", releaseToDelete, application)
-		return nil
-	}
-	idxToDelete := slices.Index(allReleases.Metadata.Releases, int64(releaseToDelete))
-
-	if idxToDelete == -1 {
-		logger.FromContext(ctx).Sugar().Infof("Could not find release '%d' for appliation '%s' to delete.", releaseToDelete, application)
-		return nil //If we don't find it, not an error, but we do nothing
-	}
-	allReleases.Metadata.Releases = append(allReleases.Metadata.Releases[:idxToDelete], allReleases.Metadata.Releases[idxToDelete+1:]...)
-	if err := h.DBInsertAllReleases(ctx, transaction, application, allReleases.Metadata.Releases, allReleases.EslVersion); err != nil {
-		return err
-	}
-	return nil
-}
-
-// DBClearReleases : Sets all_releases for app to empty list and marks every release as deleted
-func (h *DBHandler) DBClearReleases(ctx context.Context, transaction *sql.Tx, application string) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBClearReleases")
-	defer span.Finish()
-
-	allReleases, err := h.DBSelectAllReleasesOfApp(ctx, transaction, application)
-	if err != nil {
-		return err
-	}
-	if allReleases == nil {
-		logger.FromContext(ctx).Sugar().Infof("App %s does not contain any releases. No action taken", application)
-		return nil
-	}
-	for _, releaseToDelete := range allReleases.Metadata.Releases {
-		err = h.DBDeleteFromReleases(ctx, transaction, application, uint64(releaseToDelete))
-		if err != nil {
-			return err
-		}
-	}
-
-	return h.DBInsertAllReleases(ctx, transaction, application, []int64{}, allReleases.EslVersion)
-}
-
-func (h *DBHandler) DBDeleteFromReleases(ctx context.Context, transaction *sql.Tx, application string, releaseToDelete uint64) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBDeleteFromReleases")
-	defer span.Finish()
-
-	targetRelease, err := h.DBSelectReleaseByVersion(ctx, transaction, application, releaseToDelete, true)
-	if err != nil {
-		return err
-	}
-
-	if targetRelease.Deleted {
-		logger.FromContext(ctx).Sugar().Infof("Release '%d' for application '%s' has already been deleted.", releaseToDelete, application)
-		return nil
-	}
-
-	targetRelease.Deleted = true
-	if err := h.DBInsertRelease(ctx, transaction, *targetRelease); err != nil {
-		return err
-	}
-	return nil
-}
-
-func (h *DBHandler) DBInsertAllReleases(ctx context.Context, transaction *sql.Tx, app string, allVersions []int64, previousEslVersion EslVersion) error {
-	span, ctx := tracer.StartSpanFromContext(ctx, "DBInsertAllReleases")
-	defer span.Finish()
-	slices.Sort(allVersions)
-	metadataJson, err := json.Marshal(DBAllReleaseMetaData{
-		Releases: allVersions,
-	})
-	if err != nil {
-		return fmt.Errorf("insert release: could not marshal json data: %w", err)
-	}
-	insertQuery := h.AdaptQuery(
-		"INSERT INTO all_releases (eslVersion, created, appName, metadata)  VALUES (?, ?, ?, ?);",
-	)
-
-	now, err := h.DBReadTransactionTimestamp(ctx, transaction)
-	if err != nil {
-		return fmt.Errorf("DBInsertAllReleases unable to get transaction timestamp: %w", err)
-	}
-
-	span.SetTag("query", insertQuery)
-
-	_, err = transaction.Exec(
-		insertQuery,
-		previousEslVersion+1,
-		*now,
-		app,
-		metadataJson,
-	)
-	if err != nil {
-		return fmt.Errorf("could not insert all_releases for app '%s' and esl '%v' into DB. Error: %w\n", app, previousEslVersion+1, err)
-	}
-	logger.FromContext(ctx).Sugar().Infof("inserted all_releases for app '%s'", app)
-	return nil
-}
-
 // APPS
 
 func (h *DBHandler) DBWriteAllApplications(ctx context.Context, transaction *sql.Tx, previousVersion int64, applications []string) error {
@@ -5845,7 +5078,7 @@ func (h *DBHandler) RunCustomMigrationReleaseEnvironments(ctx context.Context) e
 			}
 			logger.FromContext(ctx).Sugar().Infof("updating %d releases environments", len(releasesWithoutEnvironments))
 			for _, release := range releasesWithoutEnvironments {
-				err = h.DBInsertRelease(ctx, transaction, *release)
+				err = h.DBUpdateOrCreateRelease(ctx, transaction, *release)
 				if err != nil {
 					return fmt.Errorf("could not insert release, error: %w", err)
 				}
